test(e2e): build-tag DCGM compatibility check
Keep the external package test out of the default suite and compile it explicitly in its dedicated workflow.
